+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ## v0.10.4 ##
 
 * Core: Fixed a bug with generating the correct move notation
+* Core: MoveSet reimplemented with FastAdd to improve performance
 * Engine: Fixed a bug that didn't output every valid move
 * Trainer: Fixed bug with exportai command and new version numbers
 * Trainer: Added GameString output to exceptions during battle